test_rmq/requirements.txt
pgsql
pika
PyJWT